Move InstalledPlugin to Installation
